pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "yt-tool"
version = "0.0.1"
description = "CLI + GUI download tool powered by yt-dlp"
requires-python = ">=3.10"
dependencies = [
    "yt-dlp>=2025.1.0",
]

[project.optional-dependencies]
gui = [
    "PySide6_Essentials==6.11.0",
    "shiboken6==6.11.0",
]

[project.scripts]
yt-tool = "app.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

# ---------------------------------------------------------------------------
# ruff
# ---------------------------------------------------------------------------
[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = [
    "E501",   # line too long — handled by formatter
    "SIM108", # ternary — not always clearer
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]  # assert is fine in tests

# ---------------------------------------------------------------------------
# mypy
# ---------------------------------------------------------------------------
[tool.mypy]
python_version = "3.10"
strict = false
ignore_missing_imports = true
disallow_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
exclude = ["vendor/", "build/", "dist/"]
